moved request logs to middleware and reorganized front end js
