--- controllers/showController.js
@@ -1,7 +1,0 @@
-const getAllFreePublicClaims = require('../helpers/functions/getAllFreePublicClaims.js');
-
-module.exports = {
-  getAllClaims (claimName) {
-    return getAllFreePublicClaims(claimName);
-  },
-};
