+++ public/assets/css/allStyle.css
@@ -18,15 +18,15 @@
 	margin-right: 1%;
 	padding-right: 1%;
 	border-right: 1px lightgrey solid;
-	margin-bottom: 5px;
 }
 
 .sidebar {
-	float: left;
+	float: right;
 	width: 32%;
 }
 
 footer {
+	display: inline-block;
 	width: 100%;
 	margin-bottom: 2px;
 	padding-bottom: 2px;
