+++ public/assets/js/claimPublish.js
@@ -3,81 +3,6 @@ var socket = io();
 var uploader = new SocketIOFileUpload(socket);
 var stagedFiles = null;
 
-/* helper functions */
-// create a progress animation
-function createProgressBar(element, size){ 
-	var x = 1;
-	var adder = 1;
-	function addOne(){
-		var bars = '<p>|';
-		for (var i = 0; i < x; i++){ bars += ' | '; }
-		bars += '</p>';
-		element.innerHTML = bars;
-		if (x === size){
-			adder = -1;
-		} else if ( x === 0){
-			adder = 1;
-		}
-		x += adder;
-	};
-	setInterval(addOne, 300);
-}
-// preview file and stage the image for upload
-function previewAndStageFile(selectedFile){ 
-	var preview = document.getElementById('image-preview');
-	var dropzone = document.getElementById('drop-zone');
-	var previewReader = new FileReader();
-	var nameInput = document.getElementById('publish-name'); 
-
-	preview.style.display = 'block';
-	dropzone.style.display = 'none';
-	
-	previewReader.onloadend = function () {
-		preview.src = previewReader.result;
-	};
-
-	if (selectedFile) {
-		previewReader.readAsDataURL(selectedFile); // reads the data and sets the img src
-		if (nameInput.value === "") {
-			nameInput.value = selectedFile.name.substring(0, selectedFile.name.indexOf('.'));
-		}
-		stagedFiles = [selectedFile]; // stores the selected file for upload
-	} else {
-		preview.src = '';
-	}
-}
-// update the publish status
-function updatePublishStatus(msg){
-	document.getElementById('publish-status').innerHTML = msg;
-}
-// process the drop-zone drop
-function drop_handler(ev) {
-	ev.preventDefault();
-	// if dropped items aren't files, reject them
-	var dt = ev.dataTransfer;
-	if (dt.items) {
-		if (dt.items[0].kind == 'file') {
-			var droppedFile = dt.items[0].getAsFile();
-			previewAndStageFile(droppedFile);
-		}
-	}
-}
-// prevent the browser's default drag behavior
-function dragover_handler(ev) {
-	ev.preventDefault();
-}
-// remove all of the drag data
-function dragend_handler(ev) {
-	var dt = ev.dataTransfer;
-	if (dt.items) {
-		for (var i = 0; i < dt.items.length; i++) {
-			dt.items.remove(i);
-		}
-	} else {
-		ev.dataTransfer.clearData();
-	}
-}
-
 /* configure the submit button */
 document.getElementById('publish-submit').addEventListener('click', function(event){
 	event.preventDefault();
