+++ public/assets/js/memePublish.js
@@ -7,97 +7,6 @@ var license = 'Creative Commons';
 var nsfw = false;
 var nameInput = document.getElementById("publish-name");
 
-function dataURItoBlob(dataURI) {
-    // convert base64/URLEncoded data component to raw binary data held in a string
-    var byteString;
-    if (dataURI.split(',')[0].indexOf('base64') >= 0)
-        byteString = atob(dataURI.split(',')[1]);
-    else
-        byteString = unescape(dataURI.split(',')[1]);
-
-    // separate out the mime component
-    var mimeString = dataURI.split(',')[0].split(':')[1].split(';')[0];
-
-    // write the bytes of the string to a typed array
-    var ia = new Uint8Array(byteString.length);
-    for (var i = 0; i < byteString.length; i++) {
-        ia[i] = byteString.charCodeAt(i);
-    }
-
-    return new Blob([ia], {type:mimeString});
-}
-
-function startPublish() {
-	//download the image 
-    var dataUrl = canvas.toDataURL('image/jpeg');  // canvas defined in memeDraw.js
-	var blob = dataURItoBlob(dataUrl)
-	var fileName = nameInput.value + ".jpg";  //note: need to dynamically grab type
-	var file = new File([blob], fileName, {type: 'image/jpeg', lastModified: Date.now()});
-	stageAndPublish(file); 
-};
-
-/* helper functions */
-// create a progress animation
-function createProgressBar(element, size){ 
-	var x = 1;
-	var adder = 1;
-	function addOne(){
-		var bars = '<p>|';
-		for (var i = 0; i < x; i++){ bars += ' | '; }
-		bars += '</p>';
-		element.innerHTML = bars;
-		if (x === size){
-			adder = -1;
-		} else if ( x === 0){
-			adder = 1;
-		}
-		x += adder;
-	};
-	setInterval(addOne, 300);
-}
-
-function stageAndPublish(file) {
-	var name = nameInput.value;
-	var invalidCharacters = /[^A-Za-z0-9,-]/.exec(name);
-	// validate 'name'
-	if (invalidCharacters) {
-		alert(invalidCharacters + ' is not allowed. A-Z, a-z, 0-9, "_" and "-" only.');
-		return;
-	} else if (name.length < 1) {
-		alert("You must enter a name for your claim");
-		return;
-	}
-	// stage files 
-	stagedFiles = [file]; // stores the selected file for 
-	// make sure a file was selected
-	if (stagedFiles) {
-		// make sure only 1 file was selected
-		if (stagedFiles.length < 1) {
-			alert("A file is needed");
-			return;
-		}
-		// make sure the content type is acceptable
-		switch (stagedFiles[0].type) {
-			case "image/png":
-			case "image/jpeg":
-			case "image/gif":
-			case "video/mp4":
-				uploader.submitFiles(stagedFiles);
-				break;
-			default:
-				alert("Only .png, .jpeg, .gif, and .mp4 files are currently supported");
-				break;
-		}
-	} else {
-		alert("Please select a file");
-	}
-}
-
-// update the publish status
-function updatePublishStatus(msg){
-	document.getElementById('publish-status').innerHTML = msg;
-}
-
 /* socketio-file-upload listeners */
 uploader.addEventListener('start', function(event){
 	event.file.meta.name = nameInput.value;
