+++ views/index.handlebars
@@ -9,4 +9,5 @@
 <script src="/socket.io/socket.io.js"></script>
 <script src="/siofu/client.js"></script>
 
+<script src="/assets/js/publishFunctions.js"></script>
 <script src="/assets/js/claimPublish.js"></script>
