+++ views/layouts/main.handlebars
@@ -9,8 +9,8 @@
 	<link rel="stylesheet" href="/assets/css/componentStyle.css" type="text/css">
 </head>
 <body>
-	{{{ body }}}
 	<script src="/assets/js/generalFunctions.js"></script>
+	{{{ body }}}
 	<!-- google analytics -->
 	{{ googleAnalytics }}
 </body>
