+++ views/showLite.handlebars
@@ -1,14 +1,5 @@
-<div class="wrapper">
-	{{> topBar}}
-	<div class="main">
-		{{> asset}}
-	</div>
-	<div class="sidebar">
-		<h1>{{fileInfo.name}}</h1> 
-		{{> assetLinks}}
-		{{> assetMetadata}}
-	</div>
-	{{> footer}}
+<div id="asset-placeholder" data-filepath="{{{fileInfo.filePath}}}">
+	<p> loading... </p>
 </div>
 
 <script src="/socket.io/socket.io.js"></script>
@@ -31,11 +22,11 @@
 			case 'image/gif':
 			case 'image/png':
 				const base64Image = 'data:' + data.type + ';base64,' + data.buffer;
-				document.getElementById("asset-placeholder").innerHTML = '<img class="show-asset" src="' + base64Image + '"/>';
+				document.getElementById("asset-placeholder").innerHTML = '<img class="show-asset-lite" src="' + base64Image + '"/>';
 				break;
 			case 'video/mp4':
 				const base64video = 'data:' + data.type + ';base64,' + data.buffer;
-				document.getElementById("asset-placeholder").innerHTML = '<video class="show-asset" controls> <source src="' + base64video + '"></video>';
+				document.getElementById("asset-placeholder").innerHTML = '<video class="show-asset-lite" controls> <source src="' + base64video + '"></video>';
 				break;
 			default: break;
 		}
